# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SewoApp API Documentation",
        "description": "Dokumentasi API untuk SewoApp",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Mendapatkan daftar semua kendaraan",
                "operationId": "043b3ce51f006f946511103cfe1c2ca1",
                "responses": {
                    "200": {
                        "description": "Daftar kendaraan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vehicle"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Membuat kendaraan baru",
                "operationId": "74e9ae3e5f1ee2f69f3a177820d204b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_name",
                                    "id_partner",
                                    "id_category",
                                    "price",
                                    "vehicle_year",
                                    "license_plate",
                                    "status"
                                ],
                                "properties": {
                                    "vehicle_name": {
                                        "type": "string",
                                        "example": "Honda Beat"
                                    },
                                    "id_partner": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_category": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 75000
                                    },
                                    "vehicle_year": {
                                        "type": "integer",
                                        "example": 2023
                                    },
                                    "license_plate": {
                                        "type": "string",
                                        "example": "B 1234 XYZ"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Motor matic yang nyaman untuk perjalanan"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "available"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kendaraan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Mendapatkan detail kendaraan",
                "operationId": "874afdd20b81543018ea4033fc5e914c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kendaraan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail kendaraan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kendaraan tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Mengupdate kendaraan",
                "operationId": "170b50ff07739fd3e9a2f910611ea552",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kendaraan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehicle_name": {
                                        "type": "string",
                                        "example": "Honda Beat"
                                    },
                                    "id_partner": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_category": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 75000
                                    },
                                    "vehicle_year": {
                                        "type": "integer",
                                        "example": 2023
                                    },
                                    "license_plate": {
                                        "type": "string",
                                        "example": "B 1234 XYZ"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Motor matic yang nyaman untuk perjalanan"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "available"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kendaraan berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kendaraan tidak ditemukan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Menghapus kendaraan",
                "operationId": "24688b0d004995602e93b8c1966e6676",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kendaraan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Kendaraan berhasil dihapus"
                    },
                    "404": {
                        "description": "Kendaraan tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Vehicle": {
                "required": [
                    "vehicle_name",
                    "id_partner",
                    "id_category",
                    "price",
                    "vehicle_year",
                    "license_plate",
                    "status"
                ],
                "properties": {
                    "id_vehicle": {
                        "description": "ID kendaraan",
                        "type": "integer",
                        "example": 1
                    },
                    "vehicle_name": {
                        "description": "Nama kendaraan",
                        "type": "string",
                        "example": "Honda Beat"
                    },
                    "id_partner": {
                        "description": "ID partner",
                        "type": "integer",
                        "example": 1
                    },
                    "id_category": {
                        "description": "ID kategori",
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "description": "Harga sewa per hari",
                        "type": "number",
                        "format": "float",
                        "example": 75000
                    },
                    "vehicle_year": {
                        "description": "Tahun kendaraan",
                        "type": "integer",
                        "example": 2023
                    },
                    "license_plate": {
                        "description": "Nomor plat kendaraan",
                        "type": "string",
                        "example": "B 1234 XYZ"
                    },
                    "desc": {
                        "description": "Deskripsi kendaraan",
                        "type": "string",
                        "example": "Motor matic yang nyaman"
                    },
                    "status": {
                        "description": "Status kendaraan",
                        "type": "string",
                        "example": "available"
                    },
                    "image_url": {
                        "description": "URL gambar kendaraan",
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-14T10:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}